// FrontEnd/src/app/services/api-backend/model/cita.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Clinica } from './clinica';
import { Usuario } from './usuario';
import { Caso } from './caso';
import { Receta } from './receta';
import { DiagnosticosCitum } from './diagnosticosCitum';
import { ExamenesCaso } from './examenesCaso';


export interface Cita { 
    idCita?: number;
    idCaso?: number;
    idClinica?: number;
    idUsuario?: number | null;
    fechaCita?: string;
    estado?: string | null;
    observaciones?: string | null;
    idCasoNavigation?: Caso;
    idClinicaNavigation?: Clinica;
    idUsuarioNavigation?: Usuario;
    diagnosticosCita?: Array<DiagnosticosCitum> | null;
    examenesCasos?: Array<ExamenesCaso> | null;
    receta?: Array<Receta> | null;
}