// FrontEnd/src/app/services/api-backend/model/examene.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExamenesCaso } from './examenesCaso';


export interface Examene { 
    idExamen?: number;
    nombre?: string | null;
    estado?: string | null;
    examenesCasos?: Array<ExamenesCaso> | null;
}