// FrontEnd/src/app/services/api-backend/model/medicamentosRecetum.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Producto } from './producto';
import { Receta } from './receta';


export interface MedicamentosRecetum { 
    idMedicamentosReceta?: number;
    idReceta?: number;
    idProducto?: number;
    observaciones?: string | null;
    frecuencia?: number;
    dosis?: number;
    unidadMedida?: string | null;
    idProductoNavigation?: Producto;
    idRecetaNavigation?: Receta;
}