// FrontEnd/src/app/services/api-backend/model/diagnosticosCitum.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Diagnostico } from './diagnostico';
import { Cita } from './cita';


export interface DiagnosticosCitum { 
    idDiagnosticoCita?: number;
    idCita?: number;
    idDiagnostico?: string | null;
    obresvaciones?: string | null;
    idCitaNavigation?: Cita;
    idDiagnosticoNavigation?: Diagnostico;
}